+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Send push message to the queue
+
 ## [0.0.3] - 2023-07-26
 
 ### Changed